Split lib.rs into modules
Uses `lib.rs` to declare the other modules, allowing for clearer
organization of the code base.

Drive-by:
* Removes `PartialEq` for Label impl.
